requirements.txt
mysql-connector-python>=8.0.0,<9.0.0
PyYAML>=6.0.0,<7.0.0
